requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
requests==2.31.0
